refactor: move custom sounds list query validation schema to rest-typings package
