+++ apps/meteor/app/api/server/v1/custom-sounds.ts
@@ -1,8 +1,9 @@
 import type { ICustomSound } from '@rocket.chat/core-typings';
 import { CustomSounds } from '@rocket.chat/models';
-import type { PaginatedRequest, PaginatedResult } from '@rocket.chat/rest-typings';
+import type { PaginatedResult } from '@rocket.chat/rest-typings';
 import {
 	isCustomSoundsGetOneProps,
+	isCustomSoundsListProps,
 	ajv,
 	validateBadRequestErrorResponse,
 	validateNotFoundErrorResponse,
@@ -14,38 +15,6 @@ import { escapeRegExp } from '@rocket.chat/string-helpers';
 import type { ExtractRoutesFromAPI } from '../ApiClass';
 import { API } from '../api';
 import { getPaginationItems } from '../helpers/getPaginationItems';
-
-type CustomSoundsList = PaginatedRequest<{ name?: string }>;
-
-const CustomSoundsListSchema = {
-	type: 'object',
-	properties: {
-		count: {
-			type: 'number',
-			nullable: true,
-		},
-		offset: {
-			type: 'number',
-			nullable: true,
-		},
-		sort: {
-			type: 'string',
-			nullable: true,
-		},
-		name: {
-			type: 'string',
-			nullable: true,
-		},
-		query: {
-			type: 'string',
-			nullable: true,
-		},
-	},
-	required: [],
-	additionalProperties: false,
-};
-
-export const isCustomSoundsListProps = ajv.compile<CustomSoundsList>(CustomSoundsListSchema);
 
 const customSoundsEndpoints = API.v1
 	.get(
