+++ packages/rest-typings/src/v1/customSounds.ts
@@ -1,6 +1,7 @@
 import type { ICustomSound } from '@rocket.chat/core-typings';
 
 import { ajv } from './Ajv';
+import { type PaginatedRequest } from '../helpers/PaginatedRequest';
 
 type CustomSoundsGetOne = { _id: ICustomSound['_id'] };
 
@@ -17,3 +18,35 @@ const CustomSoundsGetOneSchema = {
 };
 
 export const isCustomSoundsGetOneProps = ajv.compile<CustomSoundsGetOne>(CustomSoundsGetOneSchema);
+
+type CustomSoundsList = PaginatedRequest<{ name?: string }>;
+
+const CustomSoundsListSchema = {
+	type: 'object',
+	properties: {
+		count: {
+			type: 'number',
+			nullable: true,
+		},
+		offset: {
+			type: 'number',
+			nullable: true,
+		},
+		sort: {
+			type: 'string',
+			nullable: true,
+		},
+		name: {
+			type: 'string',
+			nullable: true,
+		},
+		query: {
+			type: 'string',
+			nullable: true,
+		},
+	},
+	required: [],
+	additionalProperties: false,
+};
+
+export const isCustomSoundsListProps = ajv.compile<CustomSoundsList>(CustomSoundsListSchema);
